Move more types to `viewModel.ts`
