+++ src/vs/editor/common/viewLayout/linesLayout.ts
@@ -3,24 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IPartialViewLinesViewportData } from 'vs/editor/common/viewLayout/viewLinesViewportData';
-import { IViewWhitespaceViewportData } from 'vs/editor/common/viewModel';
+import { IEditorWhitespace, IPartialViewLinesViewportData, IViewWhitespaceViewportData, IWhitespaceChangeAccessor } from 'vs/editor/common/viewModel';
 import * as strings from 'vs/base/common/strings';
-
-export interface IEditorWhitespace {
-	readonly id: string;
-	readonly afterLineNumber: number;
-	readonly height: number;
-}
-
-/**
- * An accessor that allows for whtiespace to be added, removed or changed in bulk.
- */
-export interface IWhitespaceChangeAccessor {
-	insertWhitespace(afterLineNumber: number, ordinal: number, heightInPx: number, minWidth: number): string;
-	changeOneWhitespace(id: string, newAfterLineNumber: number, newHeight: number): void;
-	removeWhitespace(id: string): void;
-}
 
 interface IPendingChange { id: string; newAfterLineNumber: number; newHeight: number; }
 interface IPendingRemove { id: string; }
